requirements.txt
psutil
GPUtil
pytest
mem0ai
PyYAML
